feat(40.1-02): extract CgpIntegration with hook seam and CGP defaults (D-04, D-05, D-09)
Pre-40.1, AccentApplicator owned the CodeGlance Pro reflection chain
end-to-end (resolveCgpMethods, syncCodeGlanceProViewport, the @Volatile
cached refs, and per-CGP error constants). Adding the new revert path
(revertCodeGlanceProViewport) plus the test seam (cgpRevertHook +
resetCgpRevertHookForTests) plus the project-swap entry
(syncCodeGlanceProViewportForSwap) would have pushed AccentApplicator
to 23 functions — three over the detekt TooManyFunctions threshold of
20 for objects.

Architecturally cleaner fix per CLAUDE.md "fix the actual code, never
raise thresholds": move the reflection chain and the apply / revert
workers into a dedicated `CgpIntegration` peer object. AccentApplicator
keeps only the cross-object surface that Wave 0 source-regex tests bind
by name (cgpRevertHook ThreadLocal, resetCgpRevertHookForTests,
syncCodeGlanceProViewportForSwap), plus the javap-provenance default
constants (the doc test reads them out of AccentApplicator.kt source).

Changes:
- AccentApplicator: drop CGP cached @Volatile refs and CGP error
  constants (moved to CgpIntegration); drop resolveCgpMethods +
  syncCodeGlanceProViewport (moved); add cgpRevertHook ThreadLocal +
  resetCgpRevertHookForTests test seam (Pattern I — mirror of
  ChromeDecorationsProbe.osSupplier); add three CGP_DEFAULT_VIEWPORT_*
  internal const vals with javap re-verification recipe in KDoc
  (Pattern L — D-05 source-regex regression-locked); add
  syncCodeGlanceProViewportForSwap internal wrapper (OQ1 — plan 03
  swap-path entry); update apply() call site to
  CgpIntegration.syncCodeGlanceProViewport.
- CgpIntegration (new file): owns the @Volatile cached refs,
  resolveCgpMethods, syncCodeGlanceProViewport, and
  revertCodeGlanceProViewport. The revert function checks
  AccentApplicator.cgpRevertHook BEFORE the cgpService null-guard chain
  so tests observe the three documented defaults without forcing
  reflection refs (Pattern B narrow catches —
  InvocationTargetException → ReflectiveOperationException →
  RuntimeException; never Throwable; never runCatching).

Patterns: B, I, L

Impact: AccentApplicator function count drops from 19 to 19 (net zero
after the moves). CgpIntegration owns 3 functions (well below
threshold). Wave 0 source-regex tests
AccentApplicatorCgpDefaultsDocTest, AccentApplicatorTestSeamShapeTest
flip GREEN; the integration tests targeting revertCodeGlanceProViewport
defaults via hook also pass. revertAll wiring lands in the next commit.
